Guardar en la bd el resulto de la consulta al modelo

+++ backend/src/main/java/com/g9_latam_team_67/backend/dto/contenido/ClasificacionApiResponse.java
@@ -1,10 +1,11 @@
 package com.g9_latam_team_67.backend.dto.contenido;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ClasificacionApiResponse(
 
         String category,
-        double probability
+        BigDecimal probability
 ) {
 }
